[libc] Move gettimeofday to sys/time.h header.
