+++ libc/src/time/linux/CMakeLists.txt
@@ -72,17 +72,3 @@ add_entrypoint_object(
     libc.src.__support.time.clock_settime
     libc.src.errno.errno
 )
-
-add_entrypoint_object(
-  gettimeofday
-  SRCS
-    gettimeofday.cpp
-  HDRS
-    ../gettimeofday.h
-  DEPENDS
-    libc.hdr.time_macros
-    libc.hdr.types.suseconds_t
-    libc.src.__support.time.clock_gettime
-    libc.src.__support.time.units
-    libc.src.errno.errno
-)
