+++ libc/test/src/time/CMakeLists.txt
@@ -156,18 +156,6 @@ add_libc_unittest(
 )
 
 add_libc_unittest(
-  gettimeofday_test
-  SUITE
-    libc_time_unittests
-  SRCS
-    gettimeofday_test.cpp
-  DEPENDS
-    libc.include.time
-    libc.src.time.gettimeofday
-    libc.hdr.types.struct_timeval
-)
-
-add_libc_unittest(
   gmtime_test
   SUITE
     libc_time_unittests
